fix: Python int vs bigint false positive; offer pre-commit setup in init
Python's int is arbitrary-precision, unlike a fixed-width SQL INTEGER, so
mapping it to the narrower INTEGER canonical rank produced a false "type
mismatch" warning against any dbt column typed bigint -- a very common type
for count/id columns. Now mapped to the wider BIGINT rank; a genuinely
fractional source (DECIMAL/FLOAT) is still flagged correctly.

init --interactive now also offers to set up a pre-commit hook as part of
the same wizard, instead of requiring a separate setup-precommit
invocation. The CI workflow was already generated automatically by init;
only pre-commit needed folding in.

+++ data_contract_validator/core/types.py
@@ -163,7 +163,12 @@ def normalize_sql_type(raw_type: Optional[str]) -> CanonicalType:
 
 _PYTHON_TYPE_MAP = {
     "str": CanonicalType.STRING,
-    "int": CanonicalType.INTEGER,
+    # Python's int is arbitrary-precision, unlike a fixed-width SQL INTEGER --
+    # there's no truncation risk consuming a warehouse BIGINT into it. Map to
+    # the wider rank so the numeric-widening check in types_compatible()
+    # doesn't flag e.g. a dbt `bigint` count column against `Optional[int]`
+    # as a mismatch; a genuinely fractional source (DECIMAL/FLOAT) still is.
+    "int": CanonicalType.BIGINT,
     "float": CanonicalType.FLOAT,
     "complex": CanonicalType.FLOAT,
     "bool": CanonicalType.BOOLEAN,
